fix portfolio
