+++ src/components/Skills.jsx
@@ -1,40 +1,9 @@
 import { Circle } from 'react-circle';
-import { useEffect, useReducer } from 'react';
-import axios from 'axios';
-import { skillReducer, initialState, actionTypes } from '../reducers/skillReducer';
 import { requestStates } from '../constants';
+import { useSkills } from '../customHooks/useSkills'
 
 export const Skills = () => {
-  const [state, dispatch] = useReducer(skillReducer, initialState);
-  useEffect(() => {
-    dispatch({ type: actionTypes.fetch });
-    axios.get('https://api.github.com/users/zuuumiii/repos')
-      .then((response) => {
-        const languageList = response.data.map(res => res.language);
-        const countedLanguageList = generateLanguageCountObj(languageList);
-        dispatch({ type: actionTypes.success, payload: { languageList: countedLanguageList } });
-      })
-      .catch(() => {
-        dispatch({ type: actionTypes.error });
-      });
-  }, []);
-  const generateLanguageCountObj = (allLanguageList) => {
-    const notNullLanguageList = allLanguageList.filter(language => language != null);
-    const uniqueLanguageList = [...new Set(notNullLanguageList)];
-
-    return uniqueLanguageList.map(item => {
-      return {
-        language: item,
-        count: allLanguageList.filter(language => language === item).length
-      }
-    });
-  };
-
-  const converseCountToPercentage = (count) => {
-    if (count > 10) { return 100; }
-    return count * 10;
-  };
-
+  const [sortedLnaguageList, fetchRequestState, converseCountToPercentage] = useSkills();
 
   return (
     <div id='Skills'>
@@ -44,14 +13,14 @@ export const Skills = () => {
         </div>
         <div className="skills-container">
           {
-            state.requestState === requestStates.loading && (
+            fetchRequestState === requestStates.loading && (
               <p className='description'>取得中</p>
             )
           }
 
           {
-            state.requestState === requestStates.success && (
-              state.languageList.map((item, index) => (
+            fetchRequestState === requestStates.success && (
+              sortedLnaguageList().map((item, index) => (
                 <div className='skill-item' key={index}>
                   <p className='description'><strong>{item.language}</strong></p>
                   <Circle
@@ -63,7 +32,7 @@ export const Skills = () => {
           }
 
           {
-            state.requestState === requestStates.error && (
+            fetchRequestState === requestStates.error && (
               <p className='description'>エラーが発生しました</p>
             )
           }
